safari-534.53-branch/LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.4_Comments/S7.4_A3-expected.txt

CONSOLE MESSAGE: line 82: SyntaxError: Parse error
S7.4_A3

PASS Expected parsing failure

TEST COMPLETE
